+++ pypodo/backup.py
@@ -33,7 +33,7 @@ def backup(openfile=open):
             sys.exit()
         printinfo("creating todolist backup folder")
     time_suffix = time.strftime("%Y%m%d%H%M%S")
-    todo_backup_name = ".todo" + time_suffix
+    todo_backup_name = "todo" + time_suffix
     backup_name = todobackupfolderfromconfig() + todo_backup_name
     try:
         copyfile(todofilefromconfig(), backup_name)